backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]==0.30.6
watchfiles>=0.21.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
alembic>=1.11.0

# Redis & Celery
redis>=4.5.0
celery>=5.3.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# File handling
python-multipart>=0.0.6
aiofiles>=23.0.0
requests>=2.31.0

# MinIO
minio>=7.1.0

# OpenDataLoader (HTTP 调用远程服务，无需本地安装)
# 注意：OpenDataLoader Hybrid Server 需要单独部署在 5002 端口
